+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-mobx-table
 
+## 0.0.33
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.26
+    -   @af-utils/react-table@0.0.31
+
 ## 0.0.32
 
 ### Patch Changes