+++ packages/af-utils-react-virtual-list/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @af-utils/react-virtual-list
 
+## 0.0.28
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.26
+
 ## 0.0.27
 
 ### Patch Changes